Added bytecode generator
Added the first bytecode operations and changed the to-lua to to_lua.
Also changed the Ignis data type from double to float for simplicity.

+++ analyzer/src/analyzer_value.rs
@@ -5,7 +5,7 @@ use enums::{data_type::DataType, literal_value::LiteralValue};
 pub enum AnalyzerValue {
   String(String),
   Int(i64),
-  Double(f64),
+  Float(f64),
   Boolean(bool),
   Return(Box<AnalyzerValue>),
   Function(Box<FunctionStatement>),
@@ -18,7 +18,7 @@ impl Clone for AnalyzerValue {
     match self {
       AnalyzerValue::String(s) => AnalyzerValue::String(s.clone()),
       AnalyzerValue::Int(i) => AnalyzerValue::Int(*i),
-      AnalyzerValue::Double(d) => AnalyzerValue::Double(*d),
+      AnalyzerValue::Float(d) => AnalyzerValue::Float(*d),
       AnalyzerValue::Boolean(b) => AnalyzerValue::Boolean(*b),
       AnalyzerValue::Null => AnalyzerValue::Null,
       AnalyzerValue::None => AnalyzerValue::None,
@@ -33,7 +33,7 @@ impl AnalyzerValue {
     match self {
       AnalyzerValue::String(_) => "string".to_string(),
       AnalyzerValue::Int(_) => "int".to_string(),
-      AnalyzerValue::Double(_) => "double".to_string(),
+      AnalyzerValue::Float(_) => "Float".to_string(),
       AnalyzerValue::Boolean(_) => "boolean".to_string(),
       AnalyzerValue::None | AnalyzerValue::Null => "null".to_string(),
       AnalyzerValue::Return(_) => "return".to_string(),
@@ -45,7 +45,7 @@ impl AnalyzerValue {
     match self {
       AnalyzerValue::String(_) => DataType::String,
       AnalyzerValue::Int(_) => DataType::Int,
-      AnalyzerValue::Double(_) => DataType::Double,
+      AnalyzerValue::Float(_) => DataType::Float,
       AnalyzerValue::Boolean(_) => DataType::Boolean,
       AnalyzerValue::None | AnalyzerValue::Null => DataType::None,
       AnalyzerValue::Return(r) => r.to_data_type(),
@@ -60,7 +60,7 @@ impl AnalyzerValue {
   pub fn from_literation_value(value: LiteralValue) -> Self {
     match value {
       LiteralValue::Int(i) => AnalyzerValue::Int(i),
-      LiteralValue::Double(d) => AnalyzerValue::Double(d),
+      LiteralValue::Float(d) => AnalyzerValue::Float(d),
       LiteralValue::String(s) => AnalyzerValue::String(s),
       LiteralValue::Boolean(b) => AnalyzerValue::Boolean(b),
       LiteralValue::Null => AnalyzerValue::Null,
