+++ evaluator/src/evaluator_value.rs
@@ -6,7 +6,7 @@ use crate::{callable::Callable, environment::VariableEnvironment};
 pub enum EvaluatorValue {
   String(String),
   Int(i64),
-  Double(f64),
+  Float(f64),
   Boolean(bool),
   Callable(Box<dyn Callable>),
   Return(Box<EvaluatorValue>),
@@ -25,7 +25,7 @@ impl Clone for EvaluatorValue {
     match self {
       EvaluatorValue::String(s) => EvaluatorValue::String(s.clone()),
       EvaluatorValue::Int(i) => EvaluatorValue::Int(*i),
-      EvaluatorValue::Double(d) => EvaluatorValue::Double(*d),
+      EvaluatorValue::Float(d) => EvaluatorValue::Float(*d),
       EvaluatorValue::Boolean(b) => EvaluatorValue::Boolean(*b),
       EvaluatorValue::Null => EvaluatorValue::Null,
       EvaluatorValue::None => EvaluatorValue::None,
@@ -40,7 +40,7 @@ impl EvaluatorValue {
     match self {
       EvaluatorValue::String(_) => "string".to_string(),
       EvaluatorValue::Int(_) => "int".to_string(),
-      EvaluatorValue::Double(_) => "double".to_string(),
+      EvaluatorValue::Float(_) => "Float".to_string(),
       EvaluatorValue::Boolean(_) => "boolean".to_string(),
       EvaluatorValue::None | EvaluatorValue::Null => "null".to_string(),
       EvaluatorValue::Callable(_) => "callable".to_string(),
@@ -52,7 +52,7 @@ impl EvaluatorValue {
     match self {
       EvaluatorValue::String(_) => DataType::String,
       EvaluatorValue::Int(_) => DataType::Int,
-      EvaluatorValue::Double(_) => DataType::Double,
+      EvaluatorValue::Float(_) => DataType::Float,
       EvaluatorValue::Boolean(_) => DataType::Boolean,
       EvaluatorValue::None | EvaluatorValue::Null => DataType::None,
       EvaluatorValue::Callable(callee) => callee.get_type().unwrap(),
